+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_02_20_161848) do
+ActiveRecord::Schema[8.1].define(version: 2026_03_14_000000) do
   create_table "events_events", force: :cascade do |t|
     t.integer "actor_id"
     t.string "actor_type"